+++ README.md
@@ -12,6 +12,7 @@ Parse code coverage report generated by build and post code coverage status comm
 * Jacoco ```jacoco.xml```
 * Cobertura ```cobertura.xml``` or ```cobertura-coverage.xml```
 * Clover ```clover.xml```
+* [SimpleCov JSON (Ruby)](https://github.com/vicentllongo/simplecov-json) ```coverage.json```
  
 ## How to use
 * Select or create new Jenkins build
