+++ src/main/java/com/github/terma/jenkins/githubprcoveragestatus/XmlUtils.java
@@ -15,7 +15,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 
 */
-package com.github.terma.jenkins.githubcoverageupdater;
+package com.github.terma.jenkins.githubprcoveragestatus;
 
 import org.w3c.dom.Document;
 import org.xml.sax.EntityResolver;
